+++ pmongo/__init__.py
@@ -1,5 +1,5 @@
 import document, query, utils
 
-__version__ = '1.0.9'
+__version__ = '1.0.11'
 
 __all__ = [document, query, utils]